+++ src/main/kotlin/no/nav/syfo/plugins/DependencyInjection.kt
@@ -55,7 +55,7 @@ import no.nav.syfo.narmesteleder.task.BehovMaintenanceTask
 import no.nav.syfo.pdl.PdlService
 import no.nav.syfo.pdl.client.FakePdlClient
 import no.nav.syfo.pdl.client.PdlClient
-import no.nav.syfo.pdl.leesah.PdlLeesahNameUpdateService
+import no.nav.syfo.pdl.kafka.PdlLeesahNameUpdateService
 import no.nav.syfo.person.service.PersonEnrichmentService
 import no.nav.syfo.person.task.PersonEnrichmentTask
 import no.nav.syfo.sykmelding.db.ISykmeldingDb
