+++ src/test/kotlin/no/nav/syfo/pdl/kafka/PdlLeesahNameUpdateServiceTest.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.pdl.leesah
+package no.nav.syfo.pdl.kafka
 
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.Logger
@@ -35,7 +35,7 @@ class PdlLeesahNameUpdateServiceTest :
         val pdlService = mockk<PdlService>()
         val service = PdlLeesahNameUpdateService(TestDB.exposedDatabase, pdlService)
         val logAppender = ListAppender<ILoggingEvent>()
-        val logger = LoggerFactory.getLogger("no.nav.syfo.pdl.leesah.PdlLeesahNameUpdateService") as Logger
+        val logger = LoggerFactory.getLogger(PdlLeesahNameUpdateService::class.java.name) as Logger
 
         beforeSpec {
             logger.level = Level.INFO
